Fix export crash on mesh position/tangent data and missing-image textures
Two independent crashes hit during a single model export:

1. export_model.py make_mesh_object(): position0.data and tangent0.data
   are both computed via `... @ axis_correction`, where axis_correction
   is built from a mathutils.Matrix (which iterates as plain Python
   floats, so np.array(Matrix(...)) infers float64). The matmul silently
   upcasts the float32 position/tangent arrays to float64, and
   ssbh_data_py now rejects that ("'ndarray' object cannot be converted
   to 'PyArray<T, D>'"). normal0.data already guards against this exact
   issue with an explicit .astype(np.float32) a few lines below - that
   cast was just missing on position0.data and tangent0.data. Added it
   to both, matching the existing pattern.

2. create_matl_from_blender_materials.get_texture(): crashed with
   AttributeError when a material's texture slot has no image assigned
   (sub_matl_texture.image is None) and nothing overrides it - the
   get_texture_overrides_for_bundle() dict comprehension already
   anticipated this case (it filters on `t.image is not None`), but
   get_texture() itself never handled it. Now checks for a bundle
   override first (a Shiny-style bundled material's own texture slots
   are legitimately image-less by design), then falls back to a
   per-param default texture path (matching create_default_matl_entry's
   choices) with an operator warning, instead of crashing the export.

Verified both fixes with ast-extraction tests against the real
functions/constants (dtype check for the matmul upcast; all four
get_texture cases: missing image, present image, missing image on a
different param, and override-present-despite-missing-image).

+++ source/model/export_model.py
@@ -1245,7 +1245,11 @@ def make_mesh_object(operator, context, mesh: bpy.types.Object, group_name, i, m
     positions = np.zeros(len(mesh_data.vertices) * 3, dtype=np.float32)
     mesh_data.vertices.foreach_get('co', positions)
     # The output data is flattened, so we need to reshape it into the appropriate number of rows and columns.
-    position0.data = positions.reshape((-1, 3)) @ axis_correction
+    # axis_correction is a float64 numpy array (mathutils.Matrix iterates as
+    # Python floats), so this matmul silently upcasts positions to float64 -
+    # cast back to float32 like normal0.data below does, since ssbh_data_py
+    # rejects a float64 array here.
+    position0.data = (positions.reshape((-1, 3)) @ axis_correction).astype(np.float32)
     ssbh_mesh_object.positions = [position0]
 
     # Store vertex indices as a numpy array for faster indexing later.
@@ -1410,7 +1414,8 @@ def make_mesh_object(operator, context, mesh: bpy.types.Object, group_name, i, m
 
     tangents = per_loop_to_per_vertex(loop_tangents, vertex_indices, (len(mesh.data.vertices), 3))
     bitangent_signs = per_loop_to_per_vertex(loop_bitangent_signs, vertex_indices, (len(mesh.data.vertices), 1))
-    tangent0.data = np.append(tangents @ axis_correction, bitangent_signs * -1.0, axis=1)
+    # Same float64 upcast issue as position0.data above.
+    tangent0.data = np.append(tangents @ axis_correction, bitangent_signs * -1.0, axis=1).astype(np.float32)
 
     ssbh_mesh_object.tangents = [tangent0]
